abstracted code with creation of base class
