+++ app/src/main/java/com/sschuraytz/charcoaldrawing/DrawingView.java
@@ -47,10 +47,11 @@ public class DrawingView extends View {
             case MotionEvent.ACTION_DOWN:
                 if (isEraseMode)
                 {
-                    printEraseWithLocation(pointX, pointY);
+                    printCircleWithLocation(pointX, pointY, eraseTool);
                 }
-                else {
-                    printCircleWithLocation(pointX, pointY);
+                else
+                    {
+                    printCircleWithLocation(pointX, pointY, charcoalTool);
                 }
                 previousX = pointX;
                 previousY = pointY;
@@ -58,12 +59,12 @@ public class DrawingView extends View {
             case MotionEvent.ACTION_MOVE:
                 if (isEraseMode)
                 {
-                    printEraseWithLocation(pointX, pointY);
-                    eraseContinuouslyBetweenPoints(pointX, pointY, previousX, previousY);
+                    printCircleWithLocation(pointX, pointY, eraseTool);
+                    drawContinuouslyBetweenPoints(pointX, pointY, previousX, previousY, eraseTool);
                 }
                 else {
-                    printCircleWithLocation(pointX, pointY);
-                    drawContinuouslyBetweenPoints(pointX, pointY, previousX, previousY);
+                    printCircleWithLocation(pointX, pointY, charcoalTool);
+                    drawContinuouslyBetweenPoints(pointX, pointY, previousX, previousY, charcoalTool);
                 }
                 previousX = pointX;
                 previousY = pointY;
@@ -85,7 +86,7 @@ public class DrawingView extends View {
      * @param x2
      * @param y2
      */
-    private void drawContinuouslyBetweenPoints(float x1, float y1, float x2, float y2)
+    private void drawContinuouslyBetweenPoints(float x1, float y1, float x2, float y2, Tool tool)
     {
         final float RADIUS = 10.0f;
         float dx = x2 - x1;
@@ -95,52 +96,22 @@ public class DrawingView extends View {
 
         float times = distance / RADIUS - 1;
         //ensure thin line has more dots since it doesn't have much overlap
-        float incrementer = charcoalTool.getRadius() > 5 ? 1 : 0.5f;
+        float incrementer = tool.getRadius() > 5 ? 1 : 0.5f;
         for (float i = 0; i < times; i+=incrementer)
         {
             float yIncrement = slope == 0 && dx != 0 ? 0 : dy * ( i /times);
             float xIncrement = slope == 0 ? dx * (i / times ) : yIncrement / slope;
-            bitmapCanvas.drawBitmap(charcoalTool.getBitmap(),
-                    x1 + xIncrement - charcoalTool.getRadius(),
-                    y1 + yIncrement - charcoalTool.getRadius(),
+            bitmapCanvas.drawBitmap(tool.getBitmap(),
+                    x1 + xIncrement - tool.getRadius(),
+                    y1 + yIncrement - tool.getRadius(),
                     paint);
         }
 
         if (times <= 0)
         {
-            bitmapCanvas.drawBitmap(charcoalTool.getBitmap(),
-                    x1 - charcoalTool.getRadius(),
-                    y1 - charcoalTool.getRadius(),
-                    paint);
-        }
-    }
-
-    private void eraseContinuouslyBetweenPoints(float x1, float y1, float x2, float y2)
-    {
-        final float RADIUS = 10.0f;
-        float dx = x2 - x1;
-        float dy = y2 - y1;
-        float distance = (float)Math.sqrt(dx * dx + dy * dy);
-        float slope = (dx == 0) ? 0 : dy/dx;
-
-        float times = distance / RADIUS - 1;
-        //ensure thin line has more dots since it doesn't have much overlap
-        float incrementer = eraseTool.getRadius() > 5 ? 1 : 0.5f;
-        for (float i = 0; i < times; i+=incrementer)
-        {
-            float yIncrement = slope == 0 && dx != 0 ? 0 : dy * ( i /times);
-            float xIncrement = slope == 0 ? dx * (i / times ) : yIncrement / slope;
-            bitmapCanvas.drawBitmap(eraseTool.getBitmap(),
-                    x1 + xIncrement - eraseTool.getRadius(),
-                    y1 + yIncrement - eraseTool.getRadius(),
-                    paint);
-        }
-
-        if (times <= 0)
-        {
-            bitmapCanvas.drawBitmap(eraseTool.getBitmap(),
-                    x1 - eraseTool.getRadius(),
-                    y1 - eraseTool.getRadius(),
+            bitmapCanvas.drawBitmap(tool.getBitmap(),
+                    x1 - tool.getRadius(),
+                    y1 - tool.getRadius(),
                     paint);
         }
     }
@@ -156,19 +127,11 @@ public class DrawingView extends View {
         canvas.drawBitmap(bitmap, 0, 0, paint);
     }
 
-    private void printCircleWithLocation(float pointX, float pointY)
+    private void printCircleWithLocation(float pointX, float pointY, Tool tool)
     {
-        bitmapCanvas.drawBitmap(charcoalTool.getBitmap(),
-                pointX - charcoalTool.getRadius(),
-                pointY - charcoalTool.getRadius(),
-                paint);
-    }
-
-    private void printEraseWithLocation(float pointX, float pointY)
-    {
-        bitmapCanvas.drawBitmap(eraseTool.getBitmap(),
-                pointX - eraseTool.getRadius(),
-                pointY - eraseTool.getRadius(),
+        bitmapCanvas.drawBitmap(tool.getBitmap(),
+                pointX - tool.getRadius(),
+                pointY - tool.getRadius(),
                 paint);
     }
